Added TransDecoder BLASTP confirmation of ORFs, added GT CDS as an ORF caller, updated the prepare_mikado_hints.py script

+++ eiannot/util/execute_augustus.py
@@ -8,11 +8,15 @@ import argparse
 import os
 import subprocess
 import Bio.SeqIO
+import Bio.Seq
+import Bio.SeqRecord
 import Bio.bgzf
+import pysam
 from functools import partial
 import tempfile
-import io
+import re
 import gzip
+import pybedtools
 from Mikado.parsers.GTF import GtfLine
 
 
@@ -29,11 +33,10 @@ def positive(string):
 
 def main():
 
-    fasta_index = partial(Bio.SeqIO.index, format="fasta")
-
     parser = argparse.ArgumentParser(__doc__)
-    parser.add_argument("genome", type=fasta_index)
+    parser.add_argument("genome", type=pysam.FastaFile)
     parser.add_argument("db")
+    parser.add_argument("hints")
     parser.add_argument("chunk", type=positive, help="Number of chunks to divide the genome into.")
     parser.add_argument("out")
     parser.add_argument("augustus", help="Command line to be used for this parallelisation.")
@@ -52,42 +55,51 @@ def main():
         opener = open
 
     with opener(args.out, mode="wt") as out:
+        hints = pybedtools.BedTool(args.hints).as_intervalfile()
         for chunk in session.query(Chunk).filter(Chunk.chunk_id == args.chunk):
-            chrom_file = Bio.bgzf.open(tempfile.mktemp(suffix=".fa.gz", dir=os.getcwd()), mode="wt")
+            chrom_file = Bio.bgzf.open(tempfile.mktemp(suffix=".fa.gz", dir=os.path.dirname(args.out)), mode="wt")
+            hints_file = re.sub(r"\.fa.gz", ".gff", chrom_file._handle.name)
+            # hints_file = tempfile.mktemp(suffix=".gff", dir=os.path.dirname(args.out))
             chrom, start, end = chunk.chrom, chunk.start, chunk.end
+            interval = pybedtools.Interval(chrom, start, end)
             # Using SeqIO because it is much faster than PyFaidx for this purpose
-            Bio.SeqIO.write(args.genome[chrom], chrom_file, "fasta")
+            seq = Bio.SeqRecord.SeqRecord(
+                Bio.Seq.Seq(args.genome.fetch(chrom, start, end)), description="", id=chrom
+            )
+
+            Bio.SeqIO.write(seq, chrom_file, "fasta")
+            with open(hints_file, "wt") as hint_handle:
+                for hint in hints.all_hits(interval):
+                    if hint.start < start or hint.end > end:  # Exclude partial overlaps
+                        continue
+                    hint.start -= start
+                    hint.end -= start
+                    print(hint, file=hint_handle, end='')
             chrom_file.close()
             assert os.path.exists(chrom_file._handle.name)
             assert os.stat(chrom_file._handle.name).st_size > 0
             # We will have to use bgzip to decompress on the fly
-            command = "{args.augustus} --predictionStart={start} --predictionEnd={end} "
-            command += "<(bgzip -cd {chrom_file._handle.name})"
+            command = "{args.augustus} --hintsfile={hints_file} {chrom_file._handle.name}"
             command = command.format(**locals())
             print(command)
-            aug = subprocess.Popen(command, shell=True, stdout=subprocess.PIPE, executable="/bin/bash")
-            for line in io.TextIOWrapper(aug.stdout):
+            aug = subprocess.Popen(command, shell=True, stdout=subprocess.PIPE, executable="bash",
+                                   universal_newlines=True)
+            for line in aug.stdout:
                 if line[0] == "#":
-                    print(line, file=out, end='')
+                    print(line, file=out, end="")
+                    continue
                 else:
-                    # we have to change the line so that the gene/transcript names are correct
-                    current_gene = None
                     fields = line.rstrip().split("\t")
-                    if len(fields) != 9:  # Invalid line. Continue
-                        print("#", line, file=out)
-                    elif fields[2] in ("gene", "transcript"):  # This is an invalid GTF line ...
-                        assert ";" not in fields[-1]
-                        fields[-1] = "{chrom}_{start}_{end}-{f}".format(f=fields[-1], **locals())
-                        print(*fields, file=out, sep="\t")
-                    else:
-                        line = GtfLine(line)
-                        if "transcript_id" in line.attributes:
-                            line.transcript = "{chrom}_{start}_{end}-{line.transcript}".format(**locals())
-                        if "gene_id" in line.attributes:
-                            line.gene = "{chrom}_{start}_{end}-{line.gene}".format(**locals())
-                        print(line, file=out)
+                    fields[3] = int(fields[3]) + start
+                    fields[4] = int(fields[4]) + start
+                    print(*fields, file=out, sep="\t")
 
+            aug.terminate()
+            aug.communicate()
+            if aug.returncode:
+                raise OSError("Something went wrong during the augustus run. Please inspect the logs.")
             os.remove(chrom_file._handle.name)
+            os.remove(hints_file)
 
     return
